+++ grid2op/Plot/PlotPlotly.py
@@ -62,7 +62,7 @@ except Exception as e:
 
 
 # Some utilities to plot substation, lines or get the color id for the colormap.
-def draw_sub(pos, radius=50):
+def draw_sub(pos, radius=50, line_color="LightSeaGreen"):
     """
     This function will draw the contour of a unique substation.
 
@@ -89,8 +89,8 @@ def draw_sub(pos, radius=50):
         y0=pos_y - radius,
         x1=pos_x + radius,
         y1=pos_y + radius,
-        line_color="LightSeaGreen",
-        fillcolor="lightgray"
+        line_color=line_color,
+        layer="below"
     )
     return res
 
@@ -110,20 +110,20 @@ def get_col(rho):
         The integer (between 0 and 6) of this line capacity usage in terms of color.
 
     """
-    if rho < 0.7:
+    if rho < 0.3:
         return 0
-    if rho < 0.8:
+    if rho < 0.5:
         return 1
-    if rho < 0.9:
+    if rho < 0.75:
         return 2
-    if rho < 1.:
+    if rho < 0.9:
         return 3
-    if rho < 1.1:
+    if rho < 0.95:
         return 5
     return 6
 
 
-def draw_line(pos_sub_or, pos_sub_ex, rho, color_palette, status):
+def draw_line(pos_sub_or, pos_sub_ex, rho, color_palette, status, line_color="gray"):
     """
     Draw a powerline with the color depending on its line capacity usage.
 
@@ -161,8 +161,9 @@ def draw_line(pos_sub_or, pos_sub_ex, rho, color_palette, status):
         y0=y_0,
         x1=x_1,
         y1=y_1,
+        layer="below",
         line=dict(
-            color=color_palette[get_col(rho)] if status else "gray",
+            color=line_color,
             dash=None if status else "dash"
         )
     )
@@ -242,27 +243,47 @@ class PlotPlotly(BasePlot):
 
         # define a color palette, whatever...
         sns.set()
-        # pal = sns.dark_palette("palegreen")
-        # pal = sns.color_palette("coolwarm", 7)
-        # pal = sns.light_palette("red", 7)
-        # self.cols = pal.as_hex()
         pal = sns.light_palette("darkred", 8)
         self.cols = pal.as_hex()[1:]
 
-        self.col_line = "b"
-        self.col_sub = "r"
-        self.col_load = "k"
-        self.col_gen = "g"
-        self.default_color = "k"
+        self.col_line = "royalblue"
+        self.col_sub = "red"
+        self.col_load = "black"
+        self.col_gen = "darkgreen"
+        self.default_color = "black"
 
     def init_fig(self, fig, reward, done, timestamp):
         if fig is None:
             fig = go.Figure()
         return fig
 
-    def _post_proces(self, fig, reward, done, timestamp, subs, lines, loads, gens, topos):
+    def _post_process_obs(self, fig, reward, done, timestamp, subs, lines, loads, gens, topos):
         # update the figure with all these information
-        fig.update_layout(shapes=subs + lines + loads + gens + topos)
+        traces = []
+        subs_el = []
+        lines_el = []
+        loads_el = []
+        gens_el = []
+        topos_el = []
+        for el, trace_ in subs:
+            subs_el.append(el)
+            traces.append(trace_)
+        for el, trace_ in lines:
+            lines_el.append(el)
+            traces.append(trace_)
+        for el, trace_ in loads:
+            loads_el.append(el)
+            traces.append(trace_)
+        for el, trace_ in gens:
+            gens_el.append(el)
+            traces.append(trace_)
+        for el, trace_ in topos:
+            topos_el.append(el)
+            traces.append(trace_)
+        fig.update_layout(shapes=subs_el + lines_el + loads_el + gens_el + topos_el)
+
+        for trace_ in traces:
+            fig.add_trace(trace_)
 
         # update legend, background color, size of the plot etc.
         fig.update_xaxes(range=[np.min([el for el, _ in self._layout["substations"]]) - 1.5 * (self.radius_sub +
@@ -282,18 +303,23 @@ class PlotPlotly(BasePlot):
             ),
             height=600,
             width=800,
-            plot_bgcolor="white"
+            plot_bgcolor="white",
+            yaxis={'showgrid': False, "showline": False, "zeroline": False},
+            xaxis={'showgrid': False, "showline": False, "zeroline": False}
         )
         return fig
 
     def _draw_subs_one_sub(self, fig, sub_id, center, this_col, txt_):
-        fig.add_trace(go.Scatter(x=[center.real],
-                                 y=[center.imag],
-                                 text=[txt_],
-                                 mode="text",
-                                 showlegend=False))
-        res = draw_sub(center, radius=self.radius_sub)
-        return res
+        trace = go.Scatter(x=[center[0]],
+                           y=[center[1]],
+                           text=[txt_],
+                           mode="text",
+                           showlegend=False,
+                           textfont=dict(
+                               color=this_col
+                           ))
+        res = draw_sub(center, radius=self.radius_sub, line_color=this_col)
+        return res, trace
 
     def _draw_powerlines_one_powerline(self, fig, l_id, pos_or, pos_ex, status, value, txt_, or_to_ex, this_col):
         """
@@ -312,21 +338,29 @@ class PlotPlotly(BasePlot):
                         pos_ex,
                         rho=value,
                         color_palette=self.cols,
-                        status=status)
-        fig.add_trace(go.Scatter(x=[(pos_or[0] + pos_ex[0]) / 2],
-                                 y=[(pos_or[1] + pos_ex[1]) / 2],
-                                 text=[txt_],
-                                 mode="text",
-                                 showlegend=False))
-        return tmp
+                        status=status,
+                        line_color=this_col
+                        )
+        trace = go.Scatter(x=[(pos_or[0] + pos_ex[0]) / 2],
+                           y=[(pos_or[1] + pos_ex[1]) / 2],
+                           text=[txt_],
+                           mode="text",
+                           showlegend=False,
+                           textfont=dict(
+                               color=this_col
+                           ))
+        return tmp, trace
 
     def _draw_loads_one_load(self, fig, l_id, pos_load, txt_, pos_end_line, pos_load_sub, how_center, this_col):
         # add the MW load
-        fig.add_trace(go.Scatter(x=[pos_load.real],
-                                 y=[pos_load.imag],
-                                 text=[txt_],
-                                 mode="text",
-                                 showlegend=False))
+        trace = go.Scatter(x=[pos_load.real],
+                           y=[pos_load.imag],
+                           text=[txt_],
+                           mode="text",
+                           showlegend=False,
+                           textfont=dict(
+                               color=this_col
+                           ))
         # add the line between the MW display and the substation
         # TODO later one, add something that looks like a load, a house for example
         res = go.layout.Shape(
@@ -337,18 +371,22 @@ class PlotPlotly(BasePlot):
             y0=pos_end_line.imag,
             x1=pos_load_sub[0],
             y1=pos_load_sub[1],
-            line=dict(
+            layer="below",
+            line=dict(color=this_col
             )
         )
-        return res
+        return res, trace
 
     def _draw_gens_one_gen(self, fig, g_id, pos_gen, txt_, pos_end_line, pos_gen_sub, how_center, this_col):
         # add the MW load
-        fig.add_trace(go.Scatter(x=[pos_gen.real],
-                                 y=[pos_gen.imag],
-                                 text=[txt_],
-                                 mode="text",
-                                 showlegend=False))
+        trace = go.Scatter(x=[pos_gen.real],
+                           y=[pos_gen.imag],
+                           text=[txt_],
+                           mode="text",
+                           showlegend=False,
+                           textfont=dict(
+                               color=this_col
+                           ))
         # add the line between the MW display and the substation
         # TODO later one, add something that looks like a generator, and could depend on the type of it!
         res = go.layout.Shape(
@@ -359,10 +397,11 @@ class PlotPlotly(BasePlot):
             y0=pos_end_line.imag,
             x1=pos_gen_sub[0],
             y1=pos_gen_sub[1],
-            line=dict(
+            layer="below",
+            line=dict(color=this_col
             )
         )
-        return res
+        return res, trace
 
     def _draw_topos_one_sub(self, fig, sub_id, buses_z, elements, bus_vect):
         res_sub = []
@@ -396,3 +435,6 @@ class PlotPlotly(BasePlot):
                     line=dict(color='#ff7f0e' if this_el_bus == 0 else '#1f77b4'))
                 res_sub.append(res)
         return res_sub
+
+    def _get_default_cmap(self, normalized_value):
+        return self.cols[get_col(normalized_value)]