+++ grid2op/Plot/PlotPyGame.py
@@ -177,13 +177,17 @@ class PlotPyGame(BasePlot):
                           bus_radius=bus_radius)
 
         # pygame
-        pygame.init()
+        self.__is_init = False
         self.video_width, self.video_height = 1300, 700
         self.timestep_duration_seconds = timestep_duration_seconds
-        self.display_called = False
-        self.screen = pygame.display.set_mode((self.video_width, self.video_height), pygame.RESIZABLE)
+        self.fontsize = fontsize
         self.background_color = [70, 70, 73]
-        self.font = pygame.font.Font(None, fontsize)
+
+        # init pygame
+        self.display_called = None
+        self.screen = None
+        self.font = None
+        self.init_pygame()
 
         # pause button
         self.font_pause = pygame.font.Font(None, 30)
@@ -202,6 +206,15 @@ class PlotPyGame(BasePlot):
         self.col_sub = pygame.Color(255, 0, 0)
         self.col_load = pygame.Color(0, 0, 0)
         self.col_gen = pygame.Color(0, 255, 0)
+        self.default_color = pygame.Color(0, 0, 0)
+
+    def init_pygame(self):
+        if self.__is_init is False:
+            pygame.init()
+            self.display_called = False
+            self.screen = pygame.display.set_mode((self.video_width, self.video_height), pygame.RESIZABLE)
+            self.font = pygame.font.Font(None, self.fontsize)
+            self.__is_init = True
 
     def reset(self, env):
         """
@@ -283,7 +296,10 @@ class PlotPyGame(BasePlot):
         This method is called when the renderer should be close.
         """
         self.display_called = False
-        pygame.quit()
+        try:
+            self._quit_and_close()
+        except PyGameQuit:
+            pass
 
     def _get_plot_pause(self):
         position = 300
@@ -315,6 +331,17 @@ class PlotPyGame(BasePlot):
             text_graphic = self.font.render(text_label, True, self.color_text)
             self.screen.blit(text_graphic, (self.window_grid[0]+100, 160))
 
+    def _quit_and_close(self):
+        # pygame.reset_vars()
+        # pygame.gameLoop()
+        pygame.display.quit()
+        # pygame.quit()
+        self.display_called = None
+        self.screen = None
+        self.font = None
+        self.__is_init = False
+        raise PyGameQuit()
+
     def get_rgb(self, obs, reward=None, done=None, timestamp=None):
         """
         Computes and returns the rgb 3d array from an observation, and potentially other informations.
@@ -344,6 +371,8 @@ class PlotPyGame(BasePlot):
         return pygame.surfarray.array3d(self.screen)
 
     def init_fig(self, fig, reward, done, timestamp):
+        self.init_pygame()
+
         if not self.display_called:
             self.display_called = True
             self.screen.fill(self.background_color)
@@ -355,10 +384,11 @@ class PlotPyGame(BasePlot):
             pygame.time.wait(250)  # it's in ms
 
         if has_quit:
-            raise PyGameQuit()
+            self._quit_and_close()
 
-        self.cum_reward += reward
-        self.nb_timestep += 1
+        if reward is not None:
+            self.cum_reward += reward
+            self.nb_timestep += 1
 
         # The game is not paused anymore (or never has been), I can render the next surface
         self.screen.fill(self.background_color)
@@ -423,21 +453,32 @@ class PlotPyGame(BasePlot):
                            int(self.radius_sub),
                            2)
 
+        text_graphic = self.font.render(text, True, this_col)
+        self._aligned_text(center, text_graphic, center)
+
+    def _get_default_cmap(self, normalized_val):
+        # step 0: compute thickness and color
+        max_val = 1.
+        if normalized_val < max_val:
+            amount_green = 255 - int(255. * normalized_val / max_val)
+        else:
+            amount_green = 0.
+
+        amount_red = int(255 - (50 + int(205. * normalized_val / max_val)))
+        color = pygame.Color(amount_red, amount_green, 20)
+        return color
+
     def _draw_powerlines_one_powerline(self, fig, l_id, pos_or, pos_ex, status, value, txt_, or_to_ex, this_col):
+        text_graphic = self.font.render(txt_, True, this_col)
+        pos_txt = [int((pos_or[0] + pos_ex[0]) * 0.5), int((pos_or[1] + pos_ex[1]) * 0.5)]
+        how_center = "center|center"
+        self._aligned_text(how_center, text_graphic, pos_txt)
+
         if not status:
             # line is disconnected
-            _draw_dashed_line(self.screen, pygame.Color(0, 0, 0), pos_or, pos_ex)
+            _draw_dashed_line(self.screen, this_col, pos_or, pos_ex)
         else:
             # line is connected
-
-            # step 0: compute thickness and color
-            if value < (self.rho_max / 1.5):
-                amount_green = 255 - int(255. * 1.5 * value / self.rho_max)
-            else:
-                amount_green = 0
-
-            amount_red = int(255 - (50 + int(205. * value / self.rho_max)))
-            color = pygame.Color(amount_red, amount_green, 20)
 
             width = 1
             if value > self.rho_max:
@@ -449,17 +490,21 @@ class PlotPyGame(BasePlot):
             width += 3
 
             # step 1: draw the powerline with right color and thickness
-            pygame.draw.line(self.screen, color, pos_or, pos_ex, width)
+            pygame.draw.line(self.screen, this_col, pos_or, pos_ex, width)
 
             # step 2: draw arrows indicating current flows
-            _draw_arrow(self.screen, color, pos_or, pos_ex,
+            _draw_arrow(self.screen, this_col, pos_or, pos_ex,
                         or_to_ex,
                         num_arrows=width,
                         width=width)
 
     def _aligned_text(self, pos, text_graphic, pos_text):
-        pos_x = pos_text.real
-        pos_y = pos_text.imag
+        if isinstance(pos_text, complex):
+            pos_x = pos_text.real
+            pos_y = pos_text.imag
+        else:
+            pos_x, pos_y = pos_text
+
         width = text_graphic.get_width()
         height = text_graphic.get_height()
 
@@ -473,20 +518,21 @@ class PlotPyGame(BasePlot):
             pos_y -= height // 2
         elif pos == "down|center":
             pos_x -= width // 2
+        elif pos == "center|center":
+            pos_x -= width // 2
+            pos_y -= height // 2
         self.screen.blit(text_graphic, (pos_x, pos_y))
 
     def _draw_loads_one_load(self, fig, l_id, pos_load, txt_, pos_end_line, pos_load_sub, how_center, this_col):
-        color = pygame.Color(0, 0, 0)
         width = 2
-        pygame.draw.line(self.screen, color, pos_load_sub, (pos_end_line.real, pos_end_line.imag), width)
-        text_graphic = self.font.render(txt_, True, color)
+        pygame.draw.line(self.screen, this_col, pos_load_sub, (pos_end_line.real, pos_end_line.imag), width)
+        text_graphic = self.font.render(txt_, True, this_col)
         self._aligned_text(how_center, text_graphic, pos_load)
 
     def _draw_gens_one_gen(self, fig, g_id, pos_gen, txt_, pos_end_line, pos_gen_sub, how_center, this_col):
-        color = pygame.Color(0, 0, 0)
         width = 2
-        pygame.draw.line(self.screen, color, pos_gen_sub, (pos_end_line.real, pos_end_line.imag), width)
-        text_graphic = self.font.render(txt_, True, color)
+        pygame.draw.line(self.screen, this_col, pos_gen_sub, (pos_end_line.real, pos_end_line.imag), width)
+        text_graphic = self.font.render(txt_, True, this_col)
         self._aligned_text(how_center, text_graphic, pos_gen)
         return None