+++ grid2op/Plot/__init__.py
@@ -2,13 +2,15 @@ __all__ = [
     "BasePlot", 
     "PlotMatplotlib",
     "PlotPlotly",
-    "PlotPyGame"
+    "PlotPyGame",
+    "Plotting"
 ]
 
 from grid2op.Plot.BasePlot import BasePlot
 from grid2op.Plot.PlotMatplotlib import PlotMatplotlib
 from grid2op.Plot.PlotPlotly import PlotPlotly
 from grid2op.Plot.PlotPyGame import PlotPyGame
+from grid2op.Plot.Plotting import Plotting
 
 import warnings
 
@@ -16,6 +18,6 @@ import warnings
 class PlotGraph(BasePlot):
     def __init__(self, *args, **kwargs):
         super().__init__(self, *args, **kwargs)
-        warnings.warn("TopoAndRedispAction has been renamed to TopologyAndDispatchAction"
+        warnings.warn("PlotGraph has been renamed to BasePlot"
                       " -- The old name will be removed in future versions",
                       category=PendingDeprecationWarning)